Create Canopy
I want to make a website where two or more people can red a book at the same time. So like they can see eo highlights, annotations and on what page they are. So on each page you see both of the persona highlights and if the person lefts a annotations you can click on a side start and see on what phrase the notation is left on